portable package path for bash and zsh
I've tested the sample Bash MANPAGER setting in Zsh and it works fine,
but I thought it could be made more portable by asking AppleScript for
the path to the user's Application Support folder so that they might
simply copy and paste it without customizing it.

I don't use fish, so I defer to your expertise on converting this method
to that shell.